+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - added: Log swap errors to Sentry.
 - changed: Various strings updated to UK compliance spec
+- changed: Track array of errors or AggregateErrors separately with a common tag
 - changed: Wording in light account persistent notification
 - fixed: Fix error massaging in trackError
 - fixed: Normalized error messages for tracking; removing localization from error messages.